requirements.txt
google-adk==1.19.0
google-generativeai==0.8.5
pydantic==2.12.5
mcp==1.22.0
numpy==2.3.5
pandas==2.3.3
googlesearch-python==1.3.0
typing-extensions==4.15.0